pyproject.toml
[project]
name = "mastermind-framework"
version = "1.1.0"
description = "MasterMind Framework - Cognitive architecture for expert knowledge consultation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "semver>=3.0.0",
    "pydantic>=2.0.0",
    # Phase 3: FastAPI and Web UI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.20",
    "aiofiles>=24.0.0",
    "websockets>=14.0",
    "aiosqlite>=0.22.1",
    "types-python-jose>=3.5.0.20250531",
]

[project.scripts]
mastermind = "mastermind_cli.main:cli"
mm = "mastermind_cli.main:cli"

[project.optional-dependencies]
mcp = ["notebooklm-mcp>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mastermind_cli"]

[tool.uv]
package = true

[tool.uv.workspace]
members = []

[tool.pyright]
reportUnusedParameter = "none"

[tool.mypy]
# Tier 1: Function signatures must be typed
python_version = "3.14"
disallow_untyped_defs = true  # Tier 1 - require type hints on function signatures
warn_return_any = false  # Tier 2 - enable in next phase
warn_unused_ignores = false  # Tier 3 - enable in final phase
show_error_codes = true  # Always show error codes for ignores

# Per-module overrides - strict for new/migrated code only
[[tool.mypy.overrides]]
module = "mastermind_cli.types.*"
strict = true  # New type modules are strict from day 1

[[tool.mypy.overrides]]
module = "mastermind_cli.memory.models"
strict = true  # Migrated in this plan

[[tool.mypy.overrides]]
module = "mastermind_cli.orchestrator.coordinator"
strict = true  # Migrated in this plan

[[tool.mypy.overrides]]
module = "mastermind_cli.orchestrator.mcp_wrapper"
strict = true  # Migrated in this plan

[[tool.mypy.overrides]]
module = "mastermind_cli.brains_legacy.*"
disallow_untyped_defs = false  # Legacy brains stay loose until later

# Pydantic plugin (REQUIRED for v2)
[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Pytest configuration for Phase 3
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=mastermind_cli",
    "--cov-report=term-missing:skip-covered",
]
markers = [
    "asyncio: mark test as async",
    "benchmark: mark test as performance benchmark",
    "integration: mark test as integration test",
    "e2e: mark test as end-to-end test",
]
log_cli = true
log_cli_level = "INFO"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "types-pyyaml>=6.0.12.20250915",
    # Phase 3: Testing dependencies
    "pytest-asyncio>=0.24.0",
    "pytest-benchmark>=4.0.0",
    "pytest-playwright>=0.5.0",
    "httpx-ws>=0.6.0",
    "sentence-transformers>=5.3.0",
    "scipy>=1.17.1",
]
